Use default locale instead of system
Because QLocale() returns the default locale
(which defaults to system if setDefault was not
called).

By calling system() instead of using the default
we cannot set an alternative locale (which I think
should be eventually done).

So, for now this commit does not actually change
anything but paves the road for overriding system
locale in the application settings.