Identify on mouse move, no mouse click
Adds an option to allow for identify on mouse move
without clicking on the canvas.

This option behaves almost like the default identify
option, the only difference is that the used doesn't
need to click to trigger the identify action.

A 300ms timer takes care of triggering the actual identify
to avoid performance issues when moving the mouse.